student_management_backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
mysqlclient>=2.1.0
pymysql>=1.0.0

